requirements.txt
fastapi
uvicorn
langgraph
ccxt
pandas_ta
python-dotenv
litellm
aiohttp
supabase
pydantic
